moved some functions around and wrote the ros code for the 2 new nodes
